pyproject.toml
[tool.poetry]
name = "gdbt"
version = "0.0.0"
description = "Use Infrastructure as Code to provision and manage Grafana Dashboards from Templates"
authors = ["Dmitry Kovalenko <dmytro.kovalenko@ironsrc.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
click = "^7.1.2"
rich = "^9.2.0"
attrs = "^20.3.0"
grafana-api = "^1.0.3"
requests = "^2.25.0"
jsonpath-ng = "^1.5.2"
Jinja2 = "^2.11.2"
"ruamel.yaml" = "^0.16.12"
python-box = {extras = ["ruamel.yaml"], version = "^5.2.0"}
class-registry = "^2.1.2"
deserialize = "^1.8.0"
dpath = "^2.0.1"
deepdiff = {extras = ["murmur"], version = "^5.0.2"}
python-consul2 = "^0.1.5"
halo = "^0.0.31"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.6.4"
flake8 = "^3.8.4"
mypy = "^0.790"

[tool.poetry.scripts]
gdbt = "gdbt.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
